requirements.txt
black~=22.3.0
click~=8.1.3
clipboard~=0.0.4
colorama~=0.4.4
commonmark~=0.9.1
gitdb~=4.0.9
gitpython~=3.1.27
mypy-extensions~=0.4.3
pathspec~=0.9.0
platformdirs~=2.5.2
pygments~=2.12.0
pyperclip~=1.8.2
python-dotenv~=0.20.0
python-frontmatter~=1.0.0
pyyaml~=6.0
rich~=12.4.4
setuptools~=62.3.2
smmap~=5.0.0
tomli~=2.0.1
unidecode~=1.3.4
toml~=0.10.2
